gtest: moved port and addresses test to their own test
